refactor(image-controller): remove image controller
Since image is not separate entity, delete its controller and move handling logic to user controller, to which is belongs
